requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
Flask==1.1.2
Flask-SQLAlchemy==2.4.3
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
psycopg2-binary==2.8.5
SQLAlchemy==1.3.17
urllib3==1.25.9
Werkzeug==1.0.1
pillow
dataclasses
flask_wtf
wtforms[email]
faker
